update the CTD page
